pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"

[project]
name = "auto-blog-agent"
version = "0.1.0"
description = "Multi-agent blogging pipeline that generates bilingual (KO/EN) blog posts"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.42.0",
    "langgraph>=0.2.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.12",
    "httpx>=0.27.0",
    "trafilatura>=1.12.0",
    "pymupdf>=1.24.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
    "itsdangerous>=2.1.0",
    "rich>=13.9.0",
    "youtube-transcript-api>=0.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.14.0",
    "pytest-asyncio>=0.24.0",
]

[tool.hatch.build.targets.wheel]
packages = ["config", "prompts", "agents", "parsers", "core", "web"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
